+++ src/dsp/AudioEngine.h
@@ -121,6 +121,7 @@ public:
     void setLegacyOfflineBodyDecayTimeScale (float timeScale) noexcept;
     void setLegacyOfflineHighRegisterBodyDecayTimeScale (float timeScale) noexcept;
     void setLegacyOfflineHighRegisterPartialDamping (float dampingAmount) noexcept;
+    void setLegacyOfflineHighRegisterPartialDecay (float decayAmount) noexcept;
     void setLegacyOfflineRepickCrossfadeMilliseconds (float durationMilliseconds) noexcept;
 #endif
     void setTailSustain (float newTailSustain) noexcept;